Expose selector's hierarchical level selection
